Deployment/function-source/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-dataproc==1.0.0
google-cloud-storage
protobuf